+++ packages/api/tests/es/test-es-indexer.js
@@ -13,7 +13,7 @@ const indexer = require('../../es/indexer');
 const { Search } = require('../../es/search');
 const models = require('../../models');
 const { fakeGranuleFactory, fakeCollectionFactory, deleteAliases } = require('../../lib/testUtils');
-const { constructCollectionId } = require('../../lib/utils');
+const { constructCollectionId } = require('@cumulus/common');
 const { bootstrapElasticSearch } = require('../../lambdas/bootstrap');
 const granuleSuccess = require('../data/granule_success.json');
 const granuleFailure = require('../data/granule_failed.json');
