+++ packages/common/index.js
@@ -4,7 +4,8 @@ exports.log = require('./log');
 exports.aws = require('./aws');
 exports.task = require('./task');
 exports.cliUtils = require('./cli-utils');
-exports.CollectionConfigStore = require('./collection-config-store');
+exports.constructCollectionId = require('./collection-config-store').constructCollectionId;
+exports.CollectionConfigStore = require('./collection-config-store').CollectionConfigStore;
 exports.testUtils = require('./test-utils');
 exports.FakeEarthdataLoginServer = require('./fake-earthdata-login-server');
 exports.stringUtils = require('./string');
