+++ tasks/move-granules/index.js
@@ -2,7 +2,7 @@
 
 const cumulusMessageAdapter = require('@cumulus/cumulus-message-adapter-js');
 const get = require('lodash.get');
-const { moveGranuleFile, getMetadata } = require('@cumulus/ingest/granule');
+const { moveGranuleFile, getCmrFiles, getGranuleId } = require('@cumulus/ingest/granule');
 const urljoin = require('url-join');
 const path = require('path');
 const {
@@ -12,32 +12,8 @@ const {
   }
 } = require('@cumulus/common');
 const { urlPathTemplate } = require('@cumulus/ingest/url-path-template');
-const { xmlParseOptions } = require('@cumulus/cmrjs/utils');
 const xml2js = require('xml2js');
-const { promisify } = require('util');
-
-/**
- * Parse an xml string
- *
- * @param {string} xml - xml to parse
- * @returns {Promise<Object>} promise resolves to object version of the xml
- */
-function parseXmlString(xml) {
-  return (promisify(xml2js.parseString))(xml, xmlParseOptions);
-}
-
-/**
- * Extract the granule ID from the a given s3 uri
- *
- * @param {string} uri - the s3 uri of the file
- * @param {string} regex - the regex for extracting the ID
- * @returns {string} the granule
- */
-function getGranuleId(uri, regex) {
-  const match = path.basename(uri).match(regex);
-  if (match) return match[1];
-  return null;
-}
+const log = require('@cumulus/common/log');
 
 /**
  * Creates an object with all granule files
@@ -79,37 +55,6 @@ function getAllGranules(input, granules, regex) {
   });
 
   return granulesHash;
-}
-
-/**
- * returns a list of CMR xml files
- *
- * @param {Array} input - an array of s3 uris
- * @param {string} granuleIdExtraction - a regex for extracting granule IDs
- * @returns {Promise<Array>} promise resolves to an array of objects
- * that includes CMR xmls uris and granuleIds
- */
-async function getCmrFiles(input, granuleIdExtraction) {
-  const files = [];
-  const expectedFormat = /.*\.cmr\.xml$/;
-
-  for (const filename of input) {
-    if (filename && filename.match(expectedFormat)) {
-      const metadata = await getMetadata(filename);
-      const metadataObject = await parseXmlString(metadata);
-
-      const cmrFileObject = {
-        filename,
-        metadata,
-        metadataObject,
-        granuleId: getGranuleId(filename, granuleIdExtraction)
-      };
-
-      files.push(cmrFileObject);
-    }
-  }
-
-  return files;
 }
 
 /**
@@ -178,9 +123,10 @@ async function moveGranuleFiles(granulesObject, sourceBucket) {
 
     granule.files.forEach((file) => {
       if (!(file.name.match(expectedFormat))) {
+        const fileStagingDir = file.fileStagingDir || 'file-staging';
         const source = {
           Bucket: sourceBucket,
-          Key: `${file.fileStagingDir}/${file.name}`
+          Key: `${fileStagingDir}/${file.name}`
         };
 
         const target = {
@@ -216,12 +162,11 @@ async function updateCmrFileAccessURLs(cmrFiles, granulesObject, allFiles, distE
     allFiles.forEach((file) => {
       const urlObj = {};
       if (file.bucket.type.match('protected')) {
-        console.log(`protected file: ${JSON.stringify(file)}`);
         const extension = urljoin(file.bucket.name, file.filepath);
         urlObj.URL = urljoin(distEndpoint, extension);
         urlObj.URLDescription = 'File to download';
         urls.push(urlObj);
-        console.log(`url: ${JSON.stringify(urlObj)}`);
+        log.info(`protected file: ${JSON.stringify(file)},\nurl: ${JSON.stringify(urlObj)}`);
       }
       else if (file.bucket.type.match('public')) {
         urlObj.URL = `https://${file.bucket.name}.s3.amazonaws.com/${file.filepath}`;
